configured and getAll.js
