+++ package.json
@@ -19,6 +19,7 @@
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.1",
     "nodemon": "^3.0.1",
-    "redis": "^3.0.0"
+    "redis": "^3.0.0",
+    "update": "^0.7.4"
   }
 }
